+++ fearless_simd_tests/tests/harness/ops/store_array.rs
@@ -7,10 +7,18 @@ use fearless_simd_dev_macros::simd_test;
 // One concrete test row per supported vector type.
 
 #[simd_test]
+fn store_array_mask32x4<S: Simd>(simd: S) {
+    let mask = simd.load_array::<mask32x4<S>>([-1_i32, 0_i32, -1_i32, 0_i32]);
+    let mut dest = [0_i32; 4];
+    simd.store_array(mask, &mut dest);
+    assert_eq!(dest, [-1_i32, 0_i32, -1_i32, 0_i32]);
+}
+
+#[simd_test]
 fn store_array_f32x4<S: Simd>(simd: S) {
     let a = f32x4::from_slice(simd, &[1.0, 2.0, 3.0, 4.0]);
     let mut dest = [0.0_f32; 4];
-    simd.store_array_f32x4(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, [1.0, 2.0, 3.0, 4.0]);
 }
 
@@ -18,7 +26,7 @@ fn store_array_f32x4<S: Simd>(simd: S) {
 fn store_array_f32x8<S: Simd>(simd: S) {
     let a = f32x8::from_slice(simd, &[1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0]);
     let mut dest = [0.0_f32; 8];
-    simd.store_array_f32x8(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, [1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0]);
 }
 
@@ -26,7 +34,7 @@ fn store_array_f32x8<S: Simd>(simd: S) {
 fn store_array_f64x2<S: Simd>(simd: S) {
     let a = f64x2::from_slice(simd, &[1.5, 2.5]);
     let mut dest = [0.0_f64; 2];
-    simd.store_array_f64x2(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, [1.5, 2.5]);
 }
 
@@ -34,7 +42,7 @@ fn store_array_f64x2<S: Simd>(simd: S) {
 fn store_array_f64x4<S: Simd>(simd: S) {
     let a = f64x4::from_slice(simd, &[1.5, 2.5, 3.5, 4.5]);
     let mut dest = [0.0_f64; 4];
-    simd.store_array_f64x4(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, [1.5, 2.5, 3.5, 4.5]);
 }
 
@@ -45,7 +53,7 @@ fn store_array_i8x16<S: Simd>(simd: S) {
         &[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16],
     );
     let mut dest = [0_i8; 16];
-    simd.store_array_i8x16(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(
         dest,
         [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16]
@@ -56,7 +64,7 @@ fn store_array_i8x16<S: Simd>(simd: S) {
 fn store_array_i16x8<S: Simd>(simd: S) {
     let a = i16x8::from_slice(simd, &[1, 2, 3, 4, 5, 6, 7, 8]);
     let mut dest = [0_i16; 8];
-    simd.store_array_i16x8(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, [1, 2, 3, 4, 5, 6, 7, 8]);
 }
 
@@ -64,7 +72,7 @@ fn store_array_i16x8<S: Simd>(simd: S) {
 fn store_array_i32x4<S: Simd>(simd: S) {
     let a = i32x4::from_slice(simd, &[1, 2, 3, 4]);
     let mut dest = [0_i32; 4];
-    simd.store_array_i32x4(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, [1, 2, 3, 4]);
 }
 
@@ -75,7 +83,7 @@ fn store_array_u8x16<S: Simd>(simd: S) {
         &[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16],
     );
     let mut dest = [0_u8; 16];
-    simd.store_array_u8x16(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(
         dest,
         [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16]
@@ -86,7 +94,7 @@ fn store_array_u8x16<S: Simd>(simd: S) {
 fn store_array_u16x8<S: Simd>(simd: S) {
     let a = u16x8::from_slice(simd, &[1, 2, 3, 4, 5, 6, 7, 8]);
     let mut dest = [0_u16; 8];
-    simd.store_array_u16x8(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, [1, 2, 3, 4, 5, 6, 7, 8]);
 }
 
@@ -94,7 +102,7 @@ fn store_array_u16x8<S: Simd>(simd: S) {
 fn store_array_u32x4<S: Simd>(simd: S) {
     let a = u32x4::from_slice(simd, &[1, 2, 3, 4]);
     let mut dest = [0_u32; 4];
-    simd.store_array_u32x4(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, [1, 2, 3, 4]);
 }
 
@@ -102,7 +110,7 @@ fn store_array_u32x4<S: Simd>(simd: S) {
 fn store_array_i32x8<S: Simd>(simd: S) {
     let a = i32x8::from_slice(simd, &[1, 2, 3, 4, 5, 6, 7, 8]);
     let mut dest = [0_i32; 8];
-    simd.store_array_i32x8(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, [1, 2, 3, 4, 5, 6, 7, 8]);
 }
 
@@ -110,7 +118,7 @@ fn store_array_i32x8<S: Simd>(simd: S) {
 fn store_array_u32x8<S: Simd>(simd: S) {
     let a = u32x8::from_slice(simd, &[1, 2, 3, 4, 5, 6, 7, 8]);
     let mut dest = [0_u32; 8];
-    simd.store_array_u32x8(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, [1, 2, 3, 4, 5, 6, 7, 8]);
 }
 
@@ -119,7 +127,7 @@ fn store_array_f32x16<S: Simd>(simd: S) {
     let data: [f32; 16] = core::array::from_fn(|i| i as f32);
     let a = f32x16::from_slice(simd, &data);
     let mut dest = [0.0_f32; 16];
-    simd.store_array_f32x16(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, data);
 }
 
@@ -128,7 +136,7 @@ fn store_array_f64x8<S: Simd>(simd: S) {
     let data: [f64; 8] = core::array::from_fn(|i| i as f64);
     let a = f64x8::from_slice(simd, &data);
     let mut dest = [0.0_f64; 8];
-    simd.store_array_f64x8(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, data);
 }
 
@@ -137,7 +145,7 @@ fn store_array_i8x32<S: Simd>(simd: S) {
     let data: [i8; 32] = core::array::from_fn(|i| i8::try_from(i).unwrap());
     let a = i8x32::from_slice(simd, &data);
     let mut dest = [0_i8; 32];
-    simd.store_array_i8x32(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, data);
 }
 
@@ -146,7 +154,7 @@ fn store_array_i8x64<S: Simd>(simd: S) {
     let data: [i8; 64] = core::array::from_fn(|i| i8::try_from(i).unwrap());
     let a = i8x64::from_slice(simd, &data);
     let mut dest = [0_i8; 64];
-    simd.store_array_i8x64(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, data);
 }
 
@@ -155,7 +163,7 @@ fn store_array_i16x16<S: Simd>(simd: S) {
     let data: [i16; 16] = core::array::from_fn(|i| i16::try_from(i).unwrap());
     let a = i16x16::from_slice(simd, &data);
     let mut dest = [0_i16; 16];
-    simd.store_array_i16x16(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, data);
 }
 
@@ -164,7 +172,7 @@ fn store_array_i16x32<S: Simd>(simd: S) {
     let data: [i16; 32] = core::array::from_fn(|i| i16::try_from(i).unwrap());
     let a = i16x32::from_slice(simd, &data);
     let mut dest = [0_i16; 32];
-    simd.store_array_i16x32(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, data);
 }
 
@@ -173,7 +181,7 @@ fn store_array_i32x16<S: Simd>(simd: S) {
     let data: [i32; 16] = core::array::from_fn(|i| i32::try_from(i).unwrap());
     let a = i32x16::from_slice(simd, &data);
     let mut dest = [0_i32; 16];
-    simd.store_array_i32x16(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, data);
 }
 
@@ -182,7 +190,7 @@ fn store_array_u8x32<S: Simd>(simd: S) {
     let data: [u8; 32] = core::array::from_fn(|i| u8::try_from(i).unwrap());
     let a = u8x32::from_slice(simd, &data);
     let mut dest = [0_u8; 32];
-    simd.store_array_u8x32(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, data);
 }
 
@@ -191,7 +199,7 @@ fn store_array_u8x64<S: Simd>(simd: S) {
     let data: [u8; 64] = core::array::from_fn(|i| u8::try_from(i).unwrap());
     let a = u8x64::from_slice(simd, &data);
     let mut dest = [0_u8; 64];
-    simd.store_array_u8x64(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, data);
 }
 
@@ -200,7 +208,7 @@ fn store_array_u16x16<S: Simd>(simd: S) {
     let data: [u16; 16] = core::array::from_fn(|i| u16::try_from(i).unwrap());
     let a = u16x16::from_slice(simd, &data);
     let mut dest = [0_u16; 16];
-    simd.store_array_u16x16(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, data);
 }
 
@@ -209,7 +217,7 @@ fn store_array_u16x32<S: Simd>(simd: S) {
     let data: [u16; 32] = core::array::from_fn(|i| u16::try_from(i).unwrap());
     let a = u16x32::from_slice(simd, &data);
     let mut dest = [0_u16; 32];
-    simd.store_array_u16x32(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, data);
 }
 
@@ -218,7 +226,7 @@ fn store_array_u32x16<S: Simd>(simd: S) {
     let data: [u32; 16] = core::array::from_fn(|i| u32::try_from(i).unwrap());
     let a = u32x16::from_slice(simd, &data);
     let mut dest = [0_u32; 16];
-    simd.store_array_u32x16(a, &mut dest);
+    simd.store_array(a, &mut dest);
     assert_eq!(dest, data);
 }
 
@@ -228,7 +236,7 @@ fn store_array_u32x16<S: Simd>(simd: S) {
 fn store_array_i64x2<S: Simd>(simd: S) {
     let a = i64x2::from_slice(simd, &[1_i64, -2_i64]);
     let mut out = [0_i64; 2];
-    simd.store_array_i64x2(a, &mut out);
+    simd.store_array(a, &mut out);
     assert_eq!(out, [1_i64, -2_i64]);
 }
 
@@ -236,7 +244,7 @@ fn store_array_i64x2<S: Simd>(simd: S) {
 fn store_array_i64x4<S: Simd>(simd: S) {
     let a = i64x4::from_slice(simd, &[1_i64, -2_i64, 3_i64, -4_i64]);
     let mut out = [0_i64; 4];
-    simd.store_array_i64x4(a, &mut out);
+    simd.store_array(a, &mut out);
     assert_eq!(out, [1_i64, -2_i64, 3_i64, -4_i64]);
 }
 
@@ -247,7 +255,7 @@ fn store_array_i64x8<S: Simd>(simd: S) {
         &[1_i64, -2_i64, 3_i64, -4_i64, 5_i64, -6_i64, 7_i64, -8_i64],
     );
     let mut out = [0_i64; 8];
-    simd.store_array_i64x8(a, &mut out);
+    simd.store_array(a, &mut out);
     assert_eq!(
         out,
         [1_i64, -2_i64, 3_i64, -4_i64, 5_i64, -6_i64, 7_i64, -8_i64]
@@ -258,7 +266,7 @@ fn store_array_i64x8<S: Simd>(simd: S) {
 fn store_array_u64x2<S: Simd>(simd: S) {
     let a = u64x2::from_slice(simd, &[1_u64, 2_u64]);
     let mut out = [0_u64; 2];
-    simd.store_array_u64x2(a, &mut out);
+    simd.store_array(a, &mut out);
     assert_eq!(out, [1_u64, 2_u64]);
 }
 
@@ -266,7 +274,7 @@ fn store_array_u64x2<S: Simd>(simd: S) {
 fn store_array_u64x4<S: Simd>(simd: S) {
     let a = u64x4::from_slice(simd, &[1_u64, 2_u64, 3_u64, 4_u64]);
     let mut out = [0_u64; 4];
-    simd.store_array_u64x4(a, &mut out);
+    simd.store_array(a, &mut out);
     assert_eq!(out, [1_u64, 2_u64, 3_u64, 4_u64]);
 }
 
@@ -277,7 +285,7 @@ fn store_array_u64x8<S: Simd>(simd: S) {
         &[1_u64, 2_u64, 3_u64, 4_u64, 5_u64, 6_u64, 7_u64, 8_u64],
     );
     let mut out = [0_u64; 8];
-    simd.store_array_u64x8(a, &mut out);
+    simd.store_array(a, &mut out);
     assert_eq!(
         out,
         [1_u64, 2_u64, 3_u64, 4_u64, 5_u64, 6_u64, 7_u64, 8_u64]